♻️refactor : 댓글 목록 조회 JS 분리
